+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 * [MIC-7699] CardPayment - Add common functionality of payment by card
 * [MIC-8068] SBP - Opening from main form 
 * [MIC-4650] CardsController & AddCardController for working with payment cards with own UI
+* [MIC-8027] Integrate CardsController to AddNewCard module
 
 ## [2.17.0] - 2023-02-13Z